refactor: move utility functions to a utils file
